requirements.txt
tensorflow>=2.12.0
streamlit>=1.32.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
kagglehub>=0.2.0
